Add build environment config to build for android
Took a while to figure out, but it's actually quite simple once I know how.

This commit adds a `.nix` file that installs the android linker and a version
of rust with the android stdlib. The cargo config file simply instructs cargo
to use the linker that the `.nix` file installs.