Reduce allocations made while applying patch
Did this by splitting `apply_patch` into two steps:

1. Verifying patch
2. Updating data

Before this commit, `apply_patch` would make a copy of the current state, make
changes to it, and then update the current state if there were no errors. That
meant cloning the state whether a patch was valid or not. Now the state is
never cloned.